feat: enhance Button component with class-variance-authority for variants and sizes
