+++ package.json
@@ -16,6 +16,7 @@
     "clsx": "^2.1.1",
     "@radix-ui/react-dialog": "^2.0.0",
     "@radix-ui/react-slot": "^1.0.2",
+    "class-variance-authority": "^0.7.0",
     "lucide-react": "^0.300.0",
     "next": "15.3.4",
     "prisma": "^6.10.1",
